optimized concat function with format, make format a bit more generic
